# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a API de gerenciamento de livros da Biblioteca.",
        "title": "API da Biblioteca",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/authors": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON AuthorRequest e salva o autor no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Cria um novo autor",
                "parameters": [
                    {
                        "description": "Dados do novo autor a ser criado",
                        "name": "authors",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authors.AuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Autor criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/authors.AuthorRequest"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recebe um objeto JSON AuthorRequest e atualiza o autor no banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Atualiza um autor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do autor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do Novo autor a ser atualizado",
                        "name": "authors",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authors.AuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor atualizado com sucesso"
                    },
                    "400": {
                        "description": "Requisição Inválida (JSON malformado ou campo obrigatório ausente)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui um autor específico do banco de dados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Exclui um autor pelo ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do autor a ser excluído",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Nenhum Conteúdo"
                    },
                    "400": {
                        "description": "Requisição Inválida (ID com formato incorreto)",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    },
                    "404": {
                        "description": "Livro não encontrado",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
        "/public/api/authors": {
            "get": {
                "description": "Retorna uma lista de autores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Listar autores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/authors.AuthorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
        "/public/api/authors/{id}": {
            "get": {
                "description": "Retorna um autor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "Obter autor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recebe o id do autor",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/authors.AuthorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/middleware.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "authors.AuthorRequest": {
            "description": "Dados para adicionar um autor",
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Escritor"
                },
                "name": {
                    "type": "string",
                    "example": "João Pereira"
                }
            }
        },
        "authors.AuthorResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "middleware.APIError": {
            "description": "Erro padronizado de resposta.",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}